1º functional version of the x509 identity manager
Overview:
The dojot Platform must be able to associate x.509
certificates with devices to identify them securely.
These certificates can be issued either by dojot or
by an external certification authority (functionality
not yet developed). When the certificate is issued by
dojot, it is necessary to maintain control through an
internal PKI, but when the certificate is issued
externally, dojot needs to recognize it as valid at
the user's request. To unify these two scenarios, an
x.509 Identity entity will be created (and persisted
in MongoDB), which is nothing more than an abstraction
of an x.509 certificate with some platform context
attributes. The x.509 Identity must be associated
exclusively with a device or an external application
consuming Kafka, in a 1: 1 ratio. Even after associating
an Identity with a device, it can be transferred to
another device, in which case, it will be disassociated
from the first device and then associated with the second.