1st functional version of the x509 identity management (#1469)
* 1º functional version of the x509 identity manager

Overview:
The dojot Platform must be able to associate x.509
certificates with devices to identify them securely.
These certificates can be issued either by dojot or
by an external certification authority (functionality
not yet developed). When the certificate is issued by
dojot, it is necessary to maintain control through an
internal PKI, but when the certificate is issued
externally, dojot needs to recognize it as valid at
the user's request. To unify these two scenarios, an
x.509 Identity entity will be created (and persisted
in MongoDB), which is nothing more than an abstraction
of an x.509 certificate with some platform context
attributes. The x.509 Identity must be associated
exclusively with a device or an external application
consuming Kafka, in a 1: 1 ratio. Even after associating
an Identity with a device, it can be transferred to
another device, in which case, it will be disassociated
from the first device and then associated with the second.

* fixing DeepScan issues

* fixing DeepScan issues 2

* fixing DeepScan issues 3

* fixing CodeFactor issues

* fixing CodeFactor issues 2

* fixing CodeFactor issues 3

* fixing CodeFactor issues 4

* fixing CodeFactor issues 5

* fixing CodeFactor issues 6

* fixing Codecov issues

* Upgrade to elgohr/Publish-Docker-Github-Action@2.16

* fixing Codecov issues 2

* fixing Codecov issues 3

* fixing Codecov issues 4

* fixing Codecov issues 5

* add npm run lint

* fixing Codecov issues 6

* Routes to obtain the root CA certificate and the most current CRL

* fixing Codecov issues 7

* Endpoint to issue certificate for K2V, V2K and Broker

Endpoint as a workaround so that the Devices CA can
also issue certificates for the V2K, K2V and Broker
MqTT services.
The correct would be that the Services CA to issue
these certificates, but for now, this is the solution
of lower impact.

* Securely include the tenant in the CNAME of the certificate

This implementation aims to include the tenant in the CNAME
of the certificate in a secure manner, that is, it is no
longer the user's responsibility to inform the tenant with
the device identifier in the CN field of the CSR, it is
enough for the user to enter the device identifier, the
service is responsible for include the correct tenant.

* Endpoints to obtain certificate / CRL without JWT Token and without certificate persistence (intended for applications behind the API gateway)

* Route change '/x509-certificates'->'/certificates'

The route was changed so that at API Gateway the
call to the X509 identity component services can
start with: /x509/v1/**

* fix to generate certificate when EndEntity already exists in the EJBCA database.

There was a logic problem in the script that,
since the End Entity for the NodeJS application
had already been registered in the EJBCA database,
the ${enrollmentCode} variable was not filled in
and when trying to generate a certificate (.p12)
for the NodeJS application could use, the script
could not generate this .p12 and the entire
component would become unusable.

* Separation of internal routes from external routes

Routes used by micro services have different characteristics
than routes released for use outside the API Gateway.
Open routes have more security criteria and data persistence,
but in the case of routes used by microservices, it is not
necessary to know information such as the associated Tenant,
or even to persist a certificate in MongoDB.

* fix to connect on MongoDB without user/password

* One-off changes

* Component usability documentation

* API documentation and improved file organization

* README.md for EJBCA settings

* Link between READMEs

* Link between READMEs

* README.md for Node.js App settings

* Github action to generate docs

* Changes as per comments in PR

* Changes as per comments in PR - 2

* Healthcheck treatment

* fixing Codecov issues 8

* new paging mechanism

* Changes as per comments in PR - 3

* Change to the sdk Log library

* fixing DeepScan issues 4

* Fixed Dockerfile build and component sample documentation.

* ignoring examples dir

* Very short description of the architecture

* Fix to present the image in the README

* Device Cert - Extended Key Usage: add Server Authentication

* Allow Extension Override: Subject Alternative Name

* Fix to issue many certificates per minute

Fix for Devices CA behave like a certificate factory,
therefore, it will not store the issued certificates
or the users of such certificates (End Entities), so
the Devices CA will behave like just a dummy
certificate factory.
Control will be done by the node.js application.

* README improved

* README improved

* EJBCA healthcheck does not need a certificate

* changed IP to hostname

* fix codecov flag

Co-authored-by: Mariane Previde <mprevide@cpqd.com.br>