Ticket management mechanism included
To control connections to the server via websocket,
a ticket control mechanism was implemented. Tickets
have a very short lifetime, by default 60 seconds.
The ticket must be informed as part of the URL
through the QueryString in the "ticket" field.
The ticket can be retrieved via the GET "/ticket" route
and the user must be authenticated on the platform,
in other words, it is necessary to inform the User
Access Token (JWT) in the request header to obtain
the ticket.