Feature: sdk logger with default console transport (#1537)
* Console transport enabled by default

Now the console transport is enabled by default with 'info'
logging level. This configuration is used by most of the
dojot microservices; so they don't need to worry to set it
anymore. Also, the behaviour of the method 'setTransport' changed
to replace existing configurations. This way, existing code will
continue to work.

This commit didn't change the SDK version, this should be done when
a new version of the SDK library is released.